requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile
#
aiofiles==0.5.0
    # via
    #   -r requirements.in
    #   fastapi
alembic==1.7.1
    # via -r requirements.in
aniso8601==7.0.0
    # via graphene
argon2-cffi==21.1.0
    # via -r requirements.in
async-exit-stack==1.0.1
    # via fastapi
async-generator==1.10
    # via fastapi
attrs==21.2.0
    # via pytest
backcall==0.2.0
    # via ipython
black==21.8b0
    # via -r requirements.in
certifi==2021.5.30
    # via requests
cffi==1.14.6
    # via
    #   argon2-cffi
    #   cryptography
charset-normalizer==2.0.4
    # via requests
click==7.1.2
    # via
    #   black
    #   uvicorn
cryptography==3.4.8
    # via python-jose
decorator==5.0.9
    # via ipython
dnspython==2.1.0
    # via email-validator
ecdsa==0.17.0
    # via python-jose
email-validator==1.1.3
    # via
    #   -r requirements.in
    #   fastapi
fastapi[all]==0.68.1
    # via -r requirements.in
graphene==2.1.9
    # via fastapi
graphql-core==2.3.2
    # via
    #   graphene
    #   graphql-relay
graphql-relay==2.0.1
    # via graphene
greenlet==1.1.1
    # via sqlalchemy
h11==0.12.0
    # via uvicorn
httptools==0.1.2
    # via uvicorn
idna==3.2
    # via
    #   email-validator
    #   requests
iniconfig==1.1.1
    # via pytest
ipython==7.27.0
    # via -r requirements.in
itsdangerous==1.1.0
    # via fastapi
jedi==0.18.0
    # via ipython
jinja2==2.11.3
    # via fastapi
mako==1.1.5
    # via alembic
markupsafe==2.0.1
    # via
    #   jinja2
    #   mako
matplotlib-inline==0.1.2
    # via ipython
mypy-extensions==0.4.3
    # via black
orjson==3.6.3
    # via fastapi
packaging==21.0
    # via pytest
parso==0.8.2
    # via jedi
pathspec==0.9.0
    # via black
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
platformdirs==2.3.0
    # via black
pluggy==1.0.0
    # via pytest
promise==2.3
    # via
    #   graphql-core
    #   graphql-relay
prompt-toolkit==3.0.20
    # via ipython
ptyprocess==0.7.0
    # via pexpect
py==1.10.0
    # via pytest
pyasn1==0.4.8
    # via
    #   python-jose
    #   rsa
pycparser==2.20
    # via cffi
pydantic[dotenv]==1.8.2
    # via
    #   -r requirements.in
    #   fastapi
pygments==2.10.0
    # via ipython
pyparsing==2.4.7
    # via packaging
pytest==6.2.5
    # via -r requirements.in
python-dotenv==0.19.0
    # via
    #   pydantic
    #   uvicorn
python-jose[cryptography]==3.3.0
    # via -r requirements.in
python-multipart==0.0.5
    # via fastapi
pyyaml==5.4.1
    # via
    #   fastapi
    #   uvicorn
regex==2021.8.28
    # via black
requests==2.26.0
    # via fastapi
rsa==4.7.2
    # via python-jose
rx==1.6.1
    # via graphql-core
six==1.16.0
    # via
    #   ecdsa
    #   graphene
    #   graphql-core
    #   graphql-relay
    #   python-multipart
sqlalchemy[asyncio]==1.4.23
    # via
    #   -r requirements.in
    #   alembic
starlette==0.14.2
    # via fastapi
toml==0.10.2
    # via pytest
tomli==1.2.1
    # via black
traitlets==5.1.0
    # via
    #   ipython
    #   matplotlib-inline
typing-extensions==3.10.0.2
    # via
    #   black
    #   pydantic
ujson==4.1.0
    # via fastapi
urllib3==1.26.6
    # via requests
uvicorn[standard]==0.13.4
    # via fastapi
uvloop==0.16.0
    # via uvicorn
watchgod==0.7
    # via uvicorn
wcwidth==0.2.5
    # via prompt-toolkit
websockets==8.1
    # via uvicorn
wheel==0.37.0
    # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
